fix(ffe): narrow FfeResult.serial_id from i64 to i32
The source field (ffe::Assignment.serial_id) is Option<i32>, so the value
half never needs more than 32 bits. Only has_serial_id is required to
represent the Option; the i64 width was an unnecessary widening with no
documented rationale. Narrow serial_id to i32 in the Rust struct and the
cbindgen-generated common.h ABI, and drop the `as i64` cast. The PHP
boundary still widens to zend_long in functions.c ((zend_long) promotes
int32_t cleanly), so no C reader change is needed.

Addresses review feedback from @bwoebi.

+++ components-rs/common.h
@@ -484,7 +484,7 @@ typedef struct ddog_FfeResult {
   _zend_string * allocation_key;
   int32_t reason;
   int32_t error_code;
-  int64_t serial_id;
+  int32_t serial_id;
   bool has_serial_id;
   bool do_log;
   bool valid;
